+++ pkg/util/certmanager/pki.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package pki
+package certmanager
 
 import (
 	"crypto/tls"
