--- pkg/yurthub/certificate/pki.go
@@ -1,77 +1,0 @@
-/*
-Copyright 2020 The OpenYurt Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package certificate
-
-import (
-	"crypto/tls"
-	"crypto/x509"
-	"errors"
-	"fmt"
-	"io/ioutil"
-	"os"
-
-	"k8s.io/client-go/util/certificate"
-)
-
-// GenTGenTLSConfigUseCertMgrAndCertPool generates a TLS configuration
-// using the given certificate manager and x509 CertPool
-func GenTLSConfigUseCertMgrAndCertPool(
-	m certificate.Manager,
-	root *x509.CertPool) (*tls.Config, error) {
-	tlsConfig := &tls.Config{
-		// Can't use SSLv3 because of POODLE and BEAST
-		// Can't use TLSv1.0 because of POODLE and BEAST using CBC cipher
-		// Can't use TLSv1.1 because of RC4 cipher usage
-		MinVersion: tls.VersionTLS12,
-		ClientCAs:  root,
-		ClientAuth: tls.VerifyClientCertIfGiven,
-	}
-
-	tlsConfig.GetCertificate =
-		func(*tls.ClientHelloInfo) (*tls.Certificate, error) {
-			cert := m.Current()
-			if cert == nil {
-				return &tls.Certificate{Certificate: nil}, nil
-			}
-			return cert, nil
-		}
-
-	return tlsConfig, nil
-}
-
-// GenCertPoolUseCA generates a x509 CertPool based on the given CA file
-func GenCertPoolUseCA(caFile string) (*x509.CertPool, error) {
-	if caFile == "" {
-		return nil, errors.New("CA file is not set")
-	}
-
-	if _, err := os.Stat(caFile); err != nil {
-		if os.IsNotExist(err) {
-			return nil, fmt.Errorf("CA file(%s) doesn't exist", caFile)
-		}
-		return nil, fmt.Errorf("fail to stat the CA file(%s): %s", caFile, err)
-	}
-
-	caData, err := ioutil.ReadFile(caFile)
-	if err != nil {
-		return nil, err
-	}
-
-	certPool := x509.NewCertPool()
-	certPool.AppendCertsFromPEM(caData)
-	return certPool, nil
-}
